requirements.txt
python-dotenv
fastapi
uvicorn
pymongo